#5 Rename module (core → typing)
